Add serial_sequential header and refactor serial interface
- Introduced a new `serial_sequential.h` header that aggregates all sequential interface headers, improving organization and accessibility.
- Removed the deprecated `sequential.h` header from `serial.h` to streamline header management.
- Updated several sequential interface files to include the new `internal/sequential_dispatch.h` for better encapsulation of dispatch logic.

These changes enhance the maintainability and clarity of the serial interface, aligning with recent refactoring efforts.
