+++ include/cpp_core.h
@@ -14,5 +14,6 @@
 
 #include "cpp_core/error_callback.h"
 #include "cpp_core/serial.h"
+#include "cpp_core/serial_sequential.h"
 #include "cpp_core/status_codes.h"
 #include "cpp_core/version.h"
