--- include/cpp_core/interface/sequential/sequential.h
@@ -1,18 +1,0 @@
-#pragma once
-
-// Aggregates all *Sequential headers
-#include "serial_abort_read_sequential.h"
-#include "serial_abort_write_sequential.h"
-#include "serial_clear_buffer_in_sequential.h"
-#include "serial_clear_buffer_out_sequential.h"
-#include "serial_close_sequential.h"
-#include "serial_drain_sequential.h"
-#include "serial_in_bytes_total_sequential.h"
-#include "serial_in_bytes_waiting_sequential.h"
-#include "serial_out_bytes_total_sequential.h"
-#include "serial_out_bytes_waiting_sequential.h"
-#include "serial_read_line_sequential.h"
-#include "serial_read_sequential.h"
-#include "serial_read_until_sequence_sequential.h"
-#include "serial_read_until_sequential.h"
-#include "serial_write_sequential.h"
