+++ include/cpp_core/serial.h
@@ -5,7 +5,6 @@
 
 // Aggregated interface headers
 #include "interface/get_version.h"
-#include "interface/sequential/sequential.h"
 #include "interface/serial_abort_read.h"
 #include "interface/serial_abort_write.h"
 #include "interface/serial_clear_buffer_in.h"
